2022/10/22

+++ src/main/java/com/DEAiFISH/HongxiuTeam/controller/MemberController.java
@@ -34,12 +34,13 @@ public class MemberController {
 
     /**
      * 查询所有队员
+     *
      * @param map 根据条件查询
      * @return Json格式的队员列表
      */
     @RequestMapping(value = "/member/all", method = RequestMethod.GET)
     @ResponseBody
-    public ArrayList<Member> getAllMember(@RequestBody HashMap<String,String> map) {
+    public ArrayList<Member> getAllMember(@RequestBody HashMap<String, String> map) {
         return memberService.getAllMember(map);
     }
 
@@ -52,7 +53,7 @@ public class MemberController {
      */
     @RequestMapping(value = {"/member/search"}, method = RequestMethod.GET)
     @ResponseBody
-    public ArrayList<Member> searchMember(@RequestBody HashMap<String,String> map) {
+    public ArrayList<Member> searchMember(@RequestBody HashMap<String, String> map) {
         return memberService.searchMember(map);
     }
 
@@ -60,14 +61,13 @@ public class MemberController {
     /**
      * 登录
      *
-     * @param map 信息和排序条件
+     * @param member id 和 password
      * @return
      */
     @RequestMapping(value = "/member/login", method = RequestMethod.GET)
     @ResponseBody
-    public boolean memberLogin(@RequestBody HashMap<String,String> map) {
-        ArrayList<Member> member = memberService.searchMember(map);
-        if (member.isEmpty()) {
+    public boolean memberLogin(@RequestBody Member member) {
+        if (memberService.getMemberById(member) == null) {
             return false;
         }
         return true;
@@ -77,14 +77,13 @@ public class MemberController {
     /**
      * 查询是否存在
      *
-     * @param map 根据条件查询
-     * @return true:ok;false:no;
+     * @param member Id
+     * @return true:存在;false:不存在;
      */
     @RequestMapping(value = "/member/contains", method = RequestMethod.GET)
     @ResponseBody
-    public boolean containsMemberById(@RequestBody HashMap<String,String> map) {
-        ArrayList<Member> memberList = memberService.searchMember(map);
-        if (memberList.isEmpty()) {
+    public boolean containsMemberById(@RequestBody Member member) {
+        if (memberService.getMemberById(member) == null) {
             return false;
         }
 
@@ -94,43 +93,50 @@ public class MemberController {
     /**
      * 注册队员
      *
-     * @param member 队员信息
+     * @param member id
      * @return
      */
     @RequestMapping(value = "/member/signup", method = RequestMethod.POST, produces = "application/html;charset=utf-8")
     @ResponseBody
     public String memberSignUp(@RequestBody Member member) {
-        memberService.memberSignUp(member);
-        return "注册成功";
+        if (memberService.getMemberById(member) == null) {
+            memberService.memberSignUp(member);
+            return "注册成功";
+        }
+        return "该ID已存在！注册失败!";
     }
 
     /**
      * 删除队员
      *
-     * @param member 队员信息
+     * @param member Id
      * @return
      */
     @RequestMapping(value = "/member/delete", method = RequestMethod.DELETE, produces = "application/html;charset=utf-8")
     @ResponseBody
     public String deleteMember(@RequestBody Member member) {
-
-        memberService.deleteMember(member);
-
-        return "删除成功";
+        if (memberService.getMemberById(member) != null) {
+            memberService.deleteMember(member);
+            return "删除成功";
+        }
+        return "该ID不存在！删除失败！";
     }
 
 
     /**
      * 修改成员
      *
-     * @param member 成员新信息
+     * @param member Id
      * @return
      */
     @RequestMapping(value = "/member/update", method = RequestMethod.PUT, produces = "application/html;charset=utf-8")
     @ResponseBody
     public String updateMember(@RequestBody Member member) {
-        memberService.updateMember(member);
-        return "修改成功";
+        if (memberService.getMemberById(member) != null) {
+            memberService.updateMember(member);
+            return "修改成功";
+        }
+        return "该ID不存在！修改失败！";
     }
